Separate program and USB access code.  pcsensor.c reduced to basically just main().  Added sensorapi.[ch] which abstracts all the USB access into a simple API.  pcsensor.c modified to use the new API.
