+++ src/pcsensor.c
@@ -30,35 +30,16 @@
 
 
 
-#include <usb.h>
 #include <stdio.h>
 #include <time.h>
-
+#include <ctype.h>
 #include <string.h>
-#include <errno.h>
 #include <signal.h> 
- 
- 
+
+#include "sensorapi.h"
+
 #define VERSION "1.0.0"
  
-#define VENDOR_ID  0x0c45
-#define PRODUCT_ID 0x7401
- 
-#define INTERFACE1 0x00
-#define INTERFACE2 0x01
- 
-const static int reqIntLen=8;
-const static int reqBulkLen=8;
-const static int endpoint_Int_in=0x82; /* endpoint 0x81 address for IN */
-const static int endpoint_Int_out=0x00; /* endpoint 1 address for OUT */
-const static int endpoint_Bulk_in=0x82; /* endpoint 0x81 address for IN */
-const static int endpoint_Bulk_out=0x00; /* endpoint 1 address for OUT */
-const static int timeout=5000; /* timeout in ms */
- 
-const static char uTemperatura[] = { 0x01, 0x80, 0x33, 0x01, 0x00, 0x00, 0x00, 0x00 };
-const static char uIni1[] = { 0x01, 0x82, 0x77, 0x01, 0x00, 0x00, 0x00, 0x00 };
-const static char uIni2[] = { 0x01, 0x86, 0xff, 0x01, 0x00, 0x00, 0x00, 0x00 };
-
 static int bsalir=1;
 static int debug=0;
 static int seconds=5;
@@ -66,245 +47,7 @@ static int formato=0;
 static int mrtg=0;
 static int calibration=0;
 
-
-void bad(const char *why) {
-        fprintf(stderr,"Fatal error> %s\n",why);
-        exit(17);
-}
  
- 
-usb_dev_handle *find_lvr_winusb();
- 
-void usb_detach(usb_dev_handle *lvr_winusb, int iInterface) {
-        int ret;
- 
-	ret = usb_detach_kernel_driver_np(lvr_winusb, iInterface);
-	if(ret) {
-		if(errno == ENODATA) {
-			if(debug) {
-				printf("Device already detached\n");
-			}
-		} else {
-			if(debug) {
-				printf("Detach failed: %s[%d]\n",
-				       strerror(errno), errno);
-				printf("Continuing anyway\n");
-			}
-		}
-	} else {
-		if(debug) {
-			printf("detach successful\n");
-		}
-	}
-} 
-
-usb_dev_handle* setup_libusb_access(int devicenum) {
-     usb_dev_handle *lvr_winusb;
-
-     if(debug) {
-        usb_set_debug(255);
-     } else {
-        usb_set_debug(0);
-     }
-     usb_init();
-     usb_find_busses();
-     usb_find_devices();
-             
- 
-     if(!(lvr_winusb = find_lvr_winusb(devicenum))) {
-                printf("Couldn't find the USB device, Exiting\n");
-                return NULL;
-        }
-        
-        
-        usb_detach(lvr_winusb, INTERFACE1);
-        
-
-        usb_detach(lvr_winusb, INTERFACE2);
-        
- 
-        if (usb_set_configuration(lvr_winusb, 0x01) < 0) {
-                printf("Could not set configuration 1\n");
-                return NULL;
-        }
- 
-
-        // Microdia tiene 2 interfaces
-        if (usb_claim_interface(lvr_winusb, INTERFACE1) < 0) {
-                printf("Could not claim interface\n");
-                return NULL;
-        }
- 
-        if (usb_claim_interface(lvr_winusb, INTERFACE2) < 0) {
-                printf("Could not claim interface\n");
-                return NULL;
-        }
- 
-        return lvr_winusb;
-}
- 
- 
- 
-usb_dev_handle *find_lvr_winusb(int devicenum) {
-        // iterates to the devicenum'th device for installations with multiple sensors
-        struct usb_bus *bus;
-        struct usb_device *dev;
- 
-        for (bus = usb_busses; bus; bus = bus->next) {
-        for (dev = bus->devices; dev; dev = dev->next) {
-                        if (dev->descriptor.idVendor == VENDOR_ID && 
-                                dev->descriptor.idProduct == PRODUCT_ID ) {
-                                if (devicenum>0) {
-                                  devicenum--;
-                                  continue;
-                                }
-                                usb_dev_handle *handle;
-                                if(debug) {
-                                  printf("lvr_winusb with Vendor Id: %x and Product Id: %x found.\n", VENDOR_ID, PRODUCT_ID);
-                                }
- 
-                                if (!(handle = usb_open(dev))) {
-                                        printf("Could not open USB device\n");
-                                        return NULL;
-                                }
-                                return handle;
-                        }
-                }
-        }
-        return NULL;
-}
- 
- 
-void ini_control_transfer(usb_dev_handle *dev) {
-    int r,i;
-
-    char question[] = { 0x01,0x01 };
-
-    r = usb_control_msg(dev, 0x21, 0x09, 0x0201, 0x00, (char *) question, 2, timeout);
-    if( r < 0 )
-    {
-          perror("USB control write"); bad("USB write failed"); 
-    }
-
-
-    if(debug) {
-      for (i=0;i<reqIntLen; i++) printf("%02x ",question[i] & 0xFF);
-      printf("\n");
-    }
-}
- 
-void control_transfer(usb_dev_handle *dev, const char *pquestion) {
-    int r,i;
-
-    char question[reqIntLen];
-    
-    memcpy(question, pquestion, sizeof question);
-
-    r = usb_control_msg(dev, 0x21, 0x09, 0x0200, 0x01, (char *) question, reqIntLen, timeout);
-    if( r < 0 )
-    {
-          perror("USB control write"); bad("USB write failed"); 
-    }
-
-    if(debug) {
-        for (i=0;i<reqIntLen; i++) printf("%02x ",question[i]  & 0xFF);
-        printf("\n");
-    }
-}
-
-void interrupt_transfer(usb_dev_handle *dev) {
- 
-    int r,i;
-    char answer[reqIntLen];
-    char question[reqIntLen];
-    for (i=0;i<reqIntLen; i++) question[i]=i;
-    r = usb_interrupt_write(dev, endpoint_Int_out, question, reqIntLen, timeout);
-    if( r < 0 )
-    {
-          perror("USB interrupt write"); bad("USB write failed"); 
-    }
-    r = usb_interrupt_read(dev, endpoint_Int_in, answer, reqIntLen, timeout);
-    if( r != reqIntLen )
-    {
-          perror("USB interrupt read"); bad("USB read failed"); 
-    }
-
-    if(debug) {
-       for (i=0;i<reqIntLen; i++) printf("%i, %i, \n",question[i],answer[i]);
-    }
- 
-    usb_release_interface(dev, 0);
-}
-
-void interrupt_read(usb_dev_handle *dev) {
- 
-    int r,i;
-    unsigned char answer[reqIntLen];
-    bzero(answer, reqIntLen);
-    
-    r = usb_interrupt_read(dev, 0x82, answer, reqIntLen, timeout);
-    if( r != reqIntLen )
-    {
-          perror("USB interrupt read"); bad("USB read failed"); 
-    }
-
-    if(debug) {
-       for (i=0;i<reqIntLen; i++) printf("%02x ",answer[i]  & 0xFF);
-    
-       printf("\n");
-    }
-}
-
-void interrupt_read_temperatura(usb_dev_handle *dev, float *tempC) {
- 
-    int r,i, temperature;
-    unsigned char answer[reqIntLen];
-    bzero(answer, reqIntLen);
-    
-    r = usb_interrupt_read(dev, 0x82, answer, reqIntLen, timeout);
-    if( r != reqIntLen )
-    {
-          perror("USB interrupt read"); bad("USB read failed"); 
-    }
-
-
-    if(debug) {
-      for (i=0;i<reqIntLen; i++) printf("%02x ",answer[i]  & 0xFF);
-    
-      printf("\n");
-    }
-    
-    temperature = (answer[3] & 0xFF) + (answer[2] << 8);
-    temperature += calibration;
-    *tempC = temperature * (125.0 / 32000.0);
-
-}
-
-void bulk_transfer(usb_dev_handle *dev) {
- 
-    int r,i;
-    char answer[reqBulkLen];
-
-    r = usb_bulk_write(dev, endpoint_Bulk_out, NULL, 0, timeout);
-    if( r < 0 )
-    {
-          perror("USB bulk write"); bad("USB write failed"); 
-    }
-    r = usb_bulk_read(dev, endpoint_Bulk_in, answer, reqBulkLen, timeout);
-    if( r != reqBulkLen )
-    {
-          perror("USB bulk read"); bad("USB read failed"); 
-    }
-
-
-    if(debug) {
-      for (i=0;i<reqBulkLen; i++) printf("%02x ",answer[i]  & 0xFF);
-    }
- 
-    usb_release_interface(dev, 0);
-}
- 
-
 void ex_program(int sig) {
       bsalir=1;
  
@@ -313,7 +56,8 @@ void ex_program(int sig) {
  
 int main( int argc, char **argv) {
  
-     usb_dev_handle *lvr_winusb = NULL;
+     //usb_dev_handle *lvr_winusb = NULL;
+     usb_temper_t usb_temper;
      float tempc;
      int c;
      struct tm *local;
@@ -392,30 +136,34 @@ int main( int argc, char **argv) {
         fprintf(stderr, "Non-option ARGV-elements, try -h for help.\n");
         exit(EXIT_FAILURE);
      }
+
  
-     if ((lvr_winusb = setup_libusb_access(devicenum)) == NULL) {
-         exit(EXIT_FAILURE);
-     } 
+     //if ((lvr_winusb = setup_libusb_access(devicenum)) == NULL) {
+     //    exit(EXIT_FAILURE);
+     //} 
+
+     usb_temper = usb_temper_init(devicenum);
 
      (void) signal(SIGINT, ex_program);
 
-     ini_control_transfer(lvr_winusb);
-      
-     control_transfer(lvr_winusb, uTemperatura );
-     interrupt_read(lvr_winusb);
+     //ini_control_transfer(lvr_winusb);
+     // 
+     //control_transfer(lvr_winusb, uTemperatura );
+     //interrupt_read(lvr_winusb);
  
-     control_transfer(lvr_winusb, uIni1 );
-     interrupt_read(lvr_winusb);
+     //control_transfer(lvr_winusb, uIni1 );
+     //interrupt_read(lvr_winusb);
  
-     control_transfer(lvr_winusb, uIni2 );
-     interrupt_read(lvr_winusb);
-     interrupt_read(lvr_winusb);
+     //control_transfer(lvr_winusb, uIni2 );
+     //interrupt_read(lvr_winusb);
+     //interrupt_read(lvr_winusb);
 
 
  
      do {
-           control_transfer(lvr_winusb, uTemperatura );
-           interrupt_read_temperatura(lvr_winusb, &tempc);
+           //control_transfer(lvr_winusb, uTemperatura );
+           //interrupt_read_temperatura(lvr_winusb, &tempc);
+           tempc = usb_temper_get_tempc(usb_temper);
 
            t = time(NULL);
            local = localtime(&t);
@@ -456,10 +204,12 @@ int main( int argc, char **argv) {
               sleep(seconds);
      } while (!bsalir);
                                        
-     usb_release_interface(lvr_winusb, INTERFACE1);
-     usb_release_interface(lvr_winusb, INTERFACE2);
+     //usb_release_interface(lvr_winusb, USB_TEMPER_INTERFACE1);
+     //usb_release_interface(lvr_winusb, USB_TEMPER_INTERFACE2);
      
-     usb_close(lvr_winusb); 
+     //usb_close(lvr_winusb); 
+
+     usb_temper_finish(&usb_temper);
       
      return 0; 
 }
